Chore: rename the file name and Allow USER_ROLE_OPTIONS to be passed as a prop for reusability
